Fix colour in Ticker window
